Move namespace code to subpackage
