+++ cmd/nerdctl/main.go
@@ -34,6 +34,7 @@ import (
 
 	"github.com/containerd/nerdctl/v2/cmd/nerdctl/completion"
 	"github.com/containerd/nerdctl/v2/cmd/nerdctl/helpers"
+	"github.com/containerd/nerdctl/v2/cmd/nerdctl/namespace"
 	"github.com/containerd/nerdctl/v2/cmd/nerdctl/network"
 	"github.com/containerd/nerdctl/v2/cmd/nerdctl/volume"
 	"github.com/containerd/nerdctl/v2/pkg/config"
@@ -296,7 +297,7 @@ Config file ($NERDCTL_TOML): %s
 		network.NewNetworkCommand(),
 		volume.NewVolumeCommand(),
 		newSystemCommand(),
-		newNamespaceCommand(),
+		namespace.NewNamespaceCommand(),
 		newBuilderCommand(),
 		// #endregion
